moved product scope tests into seprate module
