+++ api_tests/tests/oauth/test_jwt_unattended_access.py
@@ -1,7 +1,5 @@
 from api_tests.config_files.config import TOKEN_URL
-from api_tests.scripts.generic_helper import check_response
 import pytest
-import asyncio
 from uuid import uuid4
 from time import time
 from api_tests.config_files.config import SERVICE_NAME
@@ -14,49 +12,6 @@ from string import ascii_letters
 @pytest.mark.asyncio
 class TestJwtUnattendedAccess:
     """ A test suit to verify all the happy path oauth endpoints """
-    @pytest.fixture()
-    async def test_app_and_product(self):
-        apigee_product = ApigeeApiProducts()
-        apigee_product2 = ApigeeApiProducts()
-        await apigee_product.create_new_product()
-        await apigee_product.update_proxies([SERVICE_NAME])
-        await apigee_product2.create_new_product()
-        await apigee_product2.update_proxies([SERVICE_NAME])
-
-        apigee_app = ApigeeApiDeveloperApps()
-        await apigee_app.create_new_app(
-            callback_url="https://nhsd-apim-testing-internal-dev.herokuapp.com/callback"
-        )
-
-        # Set default JWT Testing resource url
-        await apigee_app.set_custom_attributes(
-            {
-                'jwks-resource-url': 'https://raw.githubusercontent.com/NHSDigital/'
-                                     'identity-service-jwks/main/jwks/internal-dev/'
-                                     '9baed6f4-1361-4a8e-8531-1f8426e3aba8.json'
-            }
-        )
-
-        await apigee_app.add_api_product(
-            api_products=[
-                apigee_product.name,
-                apigee_product2.name
-            ]
-        )
-
-        [await product.update_ratelimits(
-            quota=60000,
-            quota_interval="1",
-            quota_time_unit="minute",
-            rate_limit="1000ps"
-        ) for product in [apigee_product, apigee_product2]]
-
-        yield apigee_product, apigee_product2, apigee_app
-
-        await apigee_app.destroy_app()
-        await apigee_product.destroy_product()
-        await apigee_product2.destroy_product()
-
     def _update_secrets(self, request):
         if request.get("claims", None):
             if request["claims"].get("sub", None) == "/replace_me":
@@ -290,16 +245,16 @@ class TestJwtUnattendedAccess:
     ])
     @pytest.mark.apm_1521
     @pytest.mark.errors
-    async def test_invalid_jwt_claims(self, jwt_claims, expected_response, expected_status_code):
+    async def test_invalid_jwt_claims(self, jwt_claims, expected_response, expected_status_code, helper):
         self._update_secrets(jwt_claims)
         jwt = self.oauth.create_jwt(**jwt_claims)
         resp = await self.oauth.get_token_response(grant_type='client_credentials', _jwt=jwt)
 
-        assert check_response(resp, expected_status_code, expected_response)
+        assert helper.check_response(resp, expected_status_code, expected_response)
 
     @pytest.mark.apm_1521
     @pytest.mark.errors
-    async def test_reusing_same_jti(self):
+    async def test_reusing_same_jti(self, helper):
         jwt = self.oauth.create_jwt(claims={
             "sub": self.oauth.client_id,
             "iss": self.oauth.client_id,
@@ -310,11 +265,11 @@ class TestJwtUnattendedAccess:
             kid="test-1",
         )
         resp = await self.oauth.get_token_response(grant_type='client_credentials', _jwt=jwt)
-        assert check_response(
+        assert helper.check_response(
             resp, 200, ['access_token', 'expires_in', 'token_type'])
 
         resp = await self.oauth.get_token_response(grant_type='client_credentials', _jwt=jwt)
-        assert check_response(
+        assert helper.check_response(
             resp, 400, {'error': 'invalid_request', 'error_description': 'Non-unique jti claim in JWT'})
 
     @pytest.mark.happy_path
@@ -454,143 +409,3 @@ class TestJwtUnattendedAccess:
                 'error_description': 'You need to register a public key to use this '
                                      'authentication method - please contact support to configure'
             }
-
-    @pytest.mark.apm_1701
-    @pytest.mark.happy_path
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize('product_1_scopes, product_2_scopes', [
-        # Scenario 1: one product with valid scope
-        (
-            ['urn:nhsd:apim:app:level3:personal-demographics'],
-            []
-        ),
-        # Scenario 2: one product with valid scope, one product with invalid scope
-        (
-            ['urn:nhsd:apim:app:level3:personal-demographics-service'],
-            ['urn:nhsd:apim:user-nhs-id:aal3:personal-demographics-service']
-        ),
-        # Scenario 3: multiple products with valid scopes
-        (
-            ['urn:nhsd:apim:app:level3:personal-demographics-service'],
-            ['urn:nhsd:apim:app:level3:ambulance-analytics']
-        ),
-        # Scenario 4: one product with multiple valid scopes
-        (
-            ['urn:nhsd:apim:app:level3:personal-demographics', 'urn:nhsd:apim:app:level3:ambulance-analytics'],
-            []
-        ),
-        # Scenario 5: multiple products with multiple valid scopes
-        (
-            ['urn:nhsd:apim:app:level3:personal-demographics', 'urn:nhsd:apim:app:level3:ambulance-analytics'],
-            ['urn:nhsd:apim:app:level3:example-1', 'urn:nhsd:apim:app:level3:example-2']
-        ),
-        # Scenario 6: one product with multiple scopes (valid and invalid)
-        (
-            ['urn:nhsd:apim:app:level3:ambulance-analytics', 'urn:nhsd:apim:user-nhs-id:aal3:personal-demographics-service'],
-            []
-        ),
-        # Scenario 7: multiple products with multiple scopes (valid and invalid)
-        (
-            ['urn:nhsd:apim:app:level3:ambulance-analytics', 'urn:nhsd:apim:user-nhs-id:aal3:personal-demographics-service'],
-            ['urn:nhsd:apim:app:level3:example-1', 'urn:nhsd:apim:user-nhs-id:aal3:example-2']
-        ),
-        # Scenario 8: one product with valid scope with trailing and leading spaces
-        (
-            [' urn:nhsd:apim:app:level3:ambulance-analytics '],
-            []
-        ),
-    ])
-    async def test_valid_application_restricted_scope_combination(
-        self,
-        product_1_scopes,
-        product_2_scopes,
-        test_app_and_product,
-    ):
-        test_product, test_product2, test_app = test_app_and_product
-
-        await test_product.update_scopes(product_1_scopes)
-        await test_product2.update_scopes(product_2_scopes)
-
-        jwt = self.oauth.create_jwt(kid='test-1', client_id=test_app.client_id)
-        resp = await self.oauth.get_token_response(grant_type="client_credentials", _jwt=jwt)
-
-        assert list(resp['body'].keys()) == ['access_token', 'expires_in', 'token_type']
-        assert resp['status_code'] == 200
-
-    @pytest.mark.apm_1701
-    @pytest.mark.errors
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize('product_1_scopes, product_2_scopes', [
-        # Scenario 1: multiple products with no scopes
-        (
-            [],
-            []
-        ),
-        # Scenario 2: one product with invalid scope, one product with no scope
-        (
-            ['urn:nhsd:apim:user-nhs-id:aal2:personal-demographics-service'],
-            []
-        ),
-        # Scenario 3: multiple products with invalid scopes
-        (
-            ['urn:nhsd:apim:user-nhs-id:aal3:personal-demographics-service'],
-            ['urn:nhsd:apim:user-nhs-id:aal3:ambulance-analytics']
-        ),
-        # Scenario 4: one product with multiple invalid scopes
-        (
-            ['urn:nhsd:apim:user-nhs-id:aal3:personal-demographics-service', 'urn:nhsd:apim:user-nhs-id:aal3:ambulance-analytics'],
-            []
-        ),
-        # Scenario 5: multiple products with multiple invalid scopes
-        (
-            ['urn:nhsd:apim:user-nhs-id:aal3:personal-demographics-service', 'urn:nhsd:apim:user-nhs-id:aal3:ambulance-analytics'],
-            ['urn:nhsd:apim:user-nhs-id:aal3:example-1', 'urn:nhsd:apim:user-nhs-id:aal3:example-2']
-        ),
-        # Scenario 6: one product with invalid scope (wrong formation)
-        (
-            ['ThisDoesNotExist'],
-            []
-        ),
-        # Scenario 7: one product with invalid scope (special caracters)
-        (
-            ['#£$?!&%*.;@~_-'],
-            []
-        ),
-        # Scenario 8: one product with invalid scope (empty string)
-        (
-            [""],
-            []
-        ),
-        # Scenario 8: one product with invalid scope (None object)
-        (
-            [None],
-            []
-        ),
-        # Scenario 9: one product with invalid scope (missing colon)
-        (
-            ['urn:nshd:apim:app:level3personal-demographics-service'],
-            []
-        )
-    ])
-    async def test_error_application_restricted_scope_combination(
-        self,
-        product_1_scopes,
-        product_2_scopes,
-        test_app_and_product
-    ):
-        test_product, test_product2, test_app = test_app_and_product
-
-        await test_product.update_scopes(product_1_scopes)
-        await test_product2.update_scopes(product_2_scopes)
-
-        resp = await self.oauth.get_token_response(
-            grant_type="client_credentials",
-            _jwt=self.oauth.create_jwt(kid='test-1', client_id=test_app.client_id)
-        )
-
-        assert resp['status_code'] == 401
-        assert resp['body'] == {
-            "error": "unauthorized_client",
-            "error_description": "you have tried to requests authorization but your "
-                                 "application is not configured to use this authorization grant type",
-        }
